Moved initiate requests negative tests to own file
